change `sample-app` to `app`
- Removed counter/ about page
- Removed dependency on container component
- Created root-page to satisfy minimal routing
