+++ src/containers/app.html
@@ -1,0 +1,3 @@
+<div>
+  <router-outlet></router-outlet>
+</div>
