+++ src/containers/app.ts
@@ -1,0 +1,20 @@
+import {
+  Component,
+  ViewEncapsulation,
+  Inject,
+  ApplicationRef
+} from '@angular/core';
+
+import { ROUTER_DIRECTIVES } from '@angular/router';
+
+@Component({
+  selector: 'rio-app',
+  directives: [
+    ROUTER_DIRECTIVES,
+  ],
+  // Global styles imported in the app component.
+  encapsulation: ViewEncapsulation.None,
+  styles: [ require('../styles/index.css') ],
+  template: require('./app.html'),
+})
+export class RioAppComponent {};
