+++ src/containers/root-page.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'rio-root-page',
+  template: `<h1>Angular 2 + TypeScript setup</h1>`
+})
+export class RioRootPageComponent {}
