+++ src/index.html
@@ -10,7 +10,7 @@
   </head>
    <body>
     <div>
-      <rio-sample-app />
+      <rio-app />
     </div>
   </body>
 </html>
